+++ changelog.md
@@ -1,1 +1,2 @@
-- 
+- Updated forge mappings
+- Fixed Player Ghost not rendering in newer forge versions